Рефакторинг страницы настроек
