+++ src/components/Settings/Settings.css
@@ -13,4 +13,14 @@
     font-size: 36px;
     font-weight: normal;
   }
+
+  &__layout {
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-start;
+  }
+
+  &__column {
+    flex-basis: 50%;
+  }
 }
