+++ src/components/SettingsAvatar/SettingsAvatar.css
@@ -1,4 +1,4 @@
-.avatar-input {
+.settings-avatar {
   $this: &;
   $hover-overlay-opacity: 0.45;
