+++ src/components/SettingsAvatar/index.ts
@@ -1,0 +1,1 @@
+export { SettingsAvatar } from "./SettingsAvatar";
