--- src/components/SettingsForm/SettingsForm.css
@@ -1,11 +1,0 @@
-.settings-layout {
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-
-  /* Элементы */
-
-  &__column {
-    flex-basis: 50%;
-  }
-}
