--- src/components/UI/AvatarInput/index.ts
@@ -1,1 +1,0 @@
-export { AvatarInput } from "./AvatarInput";
