+++ src/services/auth/index.ts
@@ -1,3 +1,3 @@
 export { signinRequest, signupRequest, logoutRequest, fetchUserRequest } from "./actions";
 export { clearAuthError } from "./slice";
-export { isAuthenticatedSelector, errorSelector, userSelector } from "./selectors";
+export { isAuthenticatedSelector, errorSelector, userSelector, avatarSelector } from "./selectors";
